requirements.txt
streamlit
openai
python-dotenv
langchain
faiss-cpu
tiktoken
pydantic
streamlit-mic-recorder